Adds personal binary folder to dotfiles
In a hidden folder named .bin in my home directory,
often used personal scripts are saved.

Up to now, only my updateGit script is in there.
It updates my git repository, compiles an installs it.
Not listed dependencies are the needed packages for the compilation,
and an already downloaded git repository.

For installation:

	Since the downloaded git repository is already new, the
	check for new commits would state, that there is no new version.

	To force an installation anyway, provide the "--force" flag.
	It's a hacky implemantation but works for now.